sign out , move to dashboard as per auth state

+++ Firebase/authentication with email-pass/firebaseConfig.js
@@ -1,6 +1,6 @@
 import { initializeApp } from "https://www.gstatic.com/firebasejs/12.12.1/firebase-app.js";
 import { getAnalytics } from "https://www.gstatic.com/firebasejs/12.12.1/firebase-analytics.js";
-import { getAuth, createUserWithEmailAndPassword } from "https://www.gstatic.com/firebasejs/12.12.1/firebase-auth.js";
+import { signOut,getAuth,onAuthStateChanged , createUserWithEmailAndPassword } from "https://www.gstatic.com/firebasejs/12.12.1/firebase-auth.js";
 import { query, where, onSnapshot, deleteDoc, doc, setDoc, getFirestore,addDoc,collection, getDocs  } from "https://www.gstatic.com/firebasejs/12.12.1/firebase-firestore.js";
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
@@ -8,13 +8,7 @@ import { query, where, onSnapshot, deleteDoc, doc, setDoc, getFirestore,addDoc,c
 // Your web app's Firebase configuration
 // For Firebase JS SDK v7.20.0 and later, measurementId is optional
 const firebaseConfig = {
-  apiKey: "AIzaSyArV9tAE7qKNF0ub1XPKquHIzpcNYEJKvo",
-  authDomain: "ecommerce-app-3406e.firebaseapp.com",
-  projectId: "ecommerce-app-3406e",
-  storageBucket: "ecommerce-app-3406e.firebasestorage.app",
-  messagingSenderId: "602246520437",
-  appId: "1:602246520437:web:f8ab3d2edde73a4a2e623c",
-  measurementId: "G-4LRXLQR7PC"
+  ///
 };
 
 
@@ -27,4 +21,4 @@ const db = getFirestore(app);
 // Initialize Firebase Authentication and get a reference to the service
 const auth = getAuth(app);
 
-export { auth, createUserWithEmailAndPassword, onSnapshot, doc, setDoc, db, addDoc,collection, getDocs ,deleteDoc, where,query}
+export { auth,signOut,getAuth,onAuthStateChanged, createUserWithEmailAndPassword, onSnapshot, doc, setDoc, db, addDoc,collection, getDocs ,deleteDoc, where,query}
